requirements.txt
blinker==1.7.0
bson==0.5.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==2.0.7
Werkzeug==3.0.1
